+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - [mw] Don't run Ikofax expression through CQL parser
 - [ui] Fix error propagation
 - [ui] Signal search syntax in search history entry
+- [mw] Allow Ikofax syntax from URLs like ?datasource=depatisnet&query=ikofax:EP666666%2Fpn
 
 
 2018-09-28 0.152.0